+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxSeq/Internal.hs
@@ -32,9 +32,11 @@ import qualified Cardano.Crypto.Hash as Hash
 import Cardano.Ledger.Alonzo.Era
 import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsValid (..), alonzoSegwitTx)
 import Cardano.Ledger.Binary (
+  Annotated (..),
   Annotator,
   DecCBOR (..),
   EncCBORGroup (..),
+  decodeAnnotated,
   encCBOR,
   encodeFoldableEncoder,
   encodeFoldableMapEncoder,
@@ -44,14 +46,15 @@ import Cardano.Ledger.Binary (
   withSlice,
  )
 import Cardano.Ledger.Core
-import Cardano.Ledger.Shelley.BlockChain (constructMetadata)
+import Cardano.Ledger.Shelley.BlockChain (constructMetadata, indexLookupSeq)
 import Control.Monad (unless)
 import Data.ByteString (ByteString)
 import Data.ByteString.Builder (shortByteString, toLazyByteString)
 import qualified Data.ByteString.Lazy as BSL
 import Data.Coerce (coerce)
 import qualified Data.Map.Strict as Map
-import Data.Maybe.Strict (strictMaybeToMaybe)
+import Data.Maybe.Strict (maybeToStrictMaybe, strictMaybeToMaybe)
+import Data.Monoid (All (..))
 import Data.Proxy (Proxy (..))
 import qualified Data.Sequence as Seq
 import Data.Sequence.Strict (StrictSeq)
@@ -182,37 +185,37 @@ hashAlonzoTxSeq (AlonzoTxSeqRaw _ bodies ws md vs) =
 instance AlonzoEraTx era => DecCBOR (Annotator (AlonzoTxSeq era)) where
   decCBOR = do
     (bodies, bodiesAnn) <- withSlice decCBOR
-    (ws, witsAnn) <- withSlice decCBOR
-    let b = length bodies
-        inRange x = (0 <= x) && (x <= (b - 1))
-        w = length ws
+    (wits, witsAnn) <- withSlice decCBOR
+    let bodiesLength = length bodies
+        inRange x = (0 <= x) && (x <= (bodiesLength - 1))
+        witsLength = length wits
     (auxData, auxDataAnn) <- withSlice $
       do
-        m <- decCBOR
+        auxDataMap <- decCBOR
         unless
-          (all inRange (Map.keysSet m))
+          (getAll (Map.foldMapWithKey (\k _ -> All (inRange k)) auxDataMap))
           ( fail
               ( "Some Auxiliarydata index is not in the range: 0 .. "
-                  ++ show (b - 1)
+                  ++ show (bodiesLength - 1)
               )
           )
-        pure (constructMetadata b m)
+        pure (constructMetadata bodiesLength auxDataMap)
     (isValIdxs, isValAnn) <- withSlice decCBOR
-    let vs = alignedValidFlags b isValIdxs
+    let validFlags = alignedValidFlags bodiesLength isValIdxs
     unless
-      (b == w)
+      (bodiesLength == witsLength)
       ( fail $
           "different number of transaction bodies ("
-            <> show b
+            <> show bodiesLength
             <> ") and witness sets ("
-            <> show w
+            <> show witsLength
             <> ")"
       )
     unless
       (all inRange isValIdxs)
       ( fail
           ( "Some IsValid index is not in the range: 0 .. "
-              ++ show (b - 1)
+              ++ show (bodiesLength - 1)
               ++ ", "
               ++ show isValIdxs
           )
@@ -221,7 +224,7 @@ instance AlonzoEraTx era => DecCBOR (Annotator (AlonzoTxSeq era)) where
     let txns =
           sequenceA $
             StrictSeq.forceToStrict $
-              Seq.zipWith4 alonzoSegwitTx bodies ws vs auxData
+              Seq.zipWith4 alonzoSegwitTx bodies wits validFlags auxData
     pure $
       AlonzoTxSeqRaw
         <$> txns
@@ -229,6 +232,65 @@ instance AlonzoEraTx era => DecCBOR (Annotator (AlonzoTxSeq era)) where
         <*> witsAnn
         <*> auxDataAnn
         <*> isValAnn
+
+instance
+  ( AlonzoEraTx era
+  , DecCBOR (TxBody era)
+  , DecCBOR (TxWits era)
+  , DecCBOR (TxAuxData era)
+  ) =>
+  DecCBOR (AlonzoTxSeq era)
+  where
+  decCBOR = do
+    Annotated bodies bodiesBytes <- decodeAnnotated decCBOR
+    Annotated wits witsBytes <- decodeAnnotated decCBOR
+    Annotated auxDataMap auxDataBytes <- decodeAnnotated decCBOR
+    let bodiesLength = length bodies
+        inRange x = (0 <= x) && (x <= (bodiesLength - 1))
+        witsLength = length wits
+    unless
+      (getAll (Map.foldMapWithKey (\k _ -> All (inRange k)) auxDataMap))
+      ( fail
+          ( "Some Auxiliarydata index is not in the range: 0 .. "
+              ++ show (bodiesLength - 1)
+          )
+      )
+    let auxData = indexLookupSeq bodiesLength auxDataMap
+    Annotated isValidIdxs isValidBytes <- decodeAnnotated decCBOR
+    let validFlags = alignedValidFlags bodiesLength isValidIdxs
+    unless
+      (bodiesLength == witsLength)
+      ( fail $
+          "different number of transaction bodies ("
+            <> show bodiesLength
+            <> ") and witness sets ("
+            <> show witsLength
+            <> ")"
+      )
+    unless
+      (all inRange isValidIdxs)
+      ( fail
+          ( "Some IsValid index is not in the range: 0 .. "
+              ++ show (bodiesLength - 1)
+              ++ ", "
+              ++ show isValidIdxs
+          )
+      )
+    let mkTx body wit isValid aData =
+          mkBasicTx body
+            & witsTxL .~ wit
+            & auxDataTxL .~ maybeToStrictMaybe aData
+            & isValidTxL .~ isValid
+    let txs =
+          StrictSeq.forceToStrict $
+            Seq.zipWith4 mkTx bodies wits validFlags auxData
+    pure $
+      AlonzoTxSeqRaw
+        txs
+        bodiesBytes
+        witsBytes
+        auxDataBytes
+        isValidBytes
 
 --------------------------------------------------------------------------------
 -- Internal utility functions