+++ eras/shelley/impl/test/Test/Cardano/Ledger/Shelley/BinarySpec.hs
@@ -1,6 +1,10 @@
+{-# LANGUAGE TypeApplications #-}
+
 module Test.Cardano.Ledger.Shelley.BinarySpec (spec) where
 
+import Cardano.Ledger.Shelley (ShelleyEra)
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Core.Binary as Binary (decoderEquivalenceCoreEraTypesSpec)
 import qualified Test.Cardano.Ledger.Shelley.Binary.GoldenSpec as Golden
 import qualified Test.Cardano.Ledger.Shelley.Binary.RoundTripSpec as RoundTrip
 
@@ -8,3 +12,4 @@ spec :: Spec
 spec = do
   Golden.spec
   RoundTrip.spec
+  Binary.decoderEquivalenceCoreEraTypesSpec @ShelleyEra