+++ libs/cardano-protocol-tpraos/test/Test/Cardano/Protocol/Binary/CddlSpec.hs
@@ -14,11 +14,14 @@ import Cardano.Protocol.TPraos.OCert (OCert)
 import qualified Data.ByteString.Lazy as BSL
 import Test.Cardano.Ledger.Binary.Cddl (
   beforeAllCddlFile,
+  cddlDecoderEquivalenceSpec,
   cddlRoundTripAnnCborSpec,
   cddlRoundTripCborSpec,
  )
+import Test.Cardano.Ledger.Binary.Cuddle
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Shelley.Binary.Cddl (readShelleyCddlFiles)
+import Test.Cardano.Ledger.Shelley.CDDL (shelleyCDDL)
 
 spec :: Spec
 spec =
@@ -28,9 +31,16 @@ spec =
 
 specForEra :: forall era. Era era => IO [BSL.ByteString] -> Int -> Spec
 specForEra readCddlFiles n = do
-  describe (eraName @era) $
+  describe (eraName @era) $ do
+    let v = eraProtVerLow @era
     beforeAllCddlFile n readCddlFiles $ do
-      let v = eraProtVerLow @era
       cddlRoundTripAnnCborSpec @(BHeader StandardCrypto) v "header"
+      cddlRoundTripCborSpec @(BHeader StandardCrypto) v "header"
       cddlRoundTripCborSpec @(BHBody StandardCrypto) v "header_body"
       cddlRoundTripCborSpec @(CBORGroup (OCert StandardCrypto)) v "[ operational_cert ]"
+      -- TODO: add Huddle round trip tests
+      describe "DecCBOR instances equivalence via CDDL" $ do
+        cddlDecoderEquivalenceSpec @(BHeader StandardCrypto) v "header"
+    describe "DecCBOR instances equivalence via CDDL - Huddle" $ do
+      specWithHuddle shelleyCDDL 100 $ do
+        huddleDecoderEquivalenceSpec @(BHeader StandardCrypto) v "header"