+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## 1.13.0.0
 
+* Add `DecCBOR` instances for:
+  * `MaryTxBody`
+  * `TxDats`
+  * `AlonzoTxAuxData`
+  * `AlonzoScript`
+  * `Redeemers`
+  * `AlonzoTxWits`
+  * `AlonozTxBody`
+  * `AlonzoTx`
+  * `AlonzoTxSeq`
+* Remove redundant `EncCBOR (Data era)` constraint from `DecCBOR` instance for `Annotator (AlonzoTxWits era)`
 * Converted `CertState` to a type family
 * Remove `reapplyAlonzoTx` as no longer needed.
 * Add `TxInfoResult` data family, `mkTxInfoResult` and `lookupTxInfoResult` to `EraPlutusContext`
@@ -16,8 +27,11 @@
 
 ### `testlib`
 
+* Add `DecCBOR` instances for `TranslationInstance`
 * Converted `CertState` to a type family
 * Expose `alonzoFixupFees`
+* Move `Arbitrary` instances for `Data`, `BinaryData` and `Datum` to `cardano-ledger-core`
+* Move `Arbitrary` instance for `PV1.Data` to `cardano-ledger-core`
 
 ## 1.12.0.0