+++ eras/babbage/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.11.0.0
 
+* Add `DecCBOR` instance for `BabbageTxBody`
 * Converted `CertState` to a type family
 * Made the fields of predicate failures and environments lazy
 * Add `MemPack` instance for `BabbageTxOut` and `PlutusScript BabbageEra`